Extract DiscoverHeader from AsyncLoad to SSR it
